+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - removed dependency str2bool
 - use GraphParameterType for Shapes graph 
+- preserve multiple prefixes for one namespace in prefix.cc list
 
 ## [1.0.0beta1] 2024-07-24